requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
umap-learn
jupyterlab
ipykernel
scipy
hdbscan